requirements.txt
aiogram==2.25.1
loguru==0.7.0
openpyxl==3.1.2
peewee==3.16.2
python-dotenv==1.0.0
Flask==3.0.3
requests==2.32.3
